+++ test/service/common.dart
@@ -13,7 +13,8 @@ class Todo extends Model {
 
 class TodoService extends MapService {
   TodoService() : super() {
-    configuration['ws:filter'] = (HookedServiceEvent e, WebSocketContext socket) {
+    configuration['ws:filter'] =
+        (HookedServiceEvent e, WebSocketContext socket) {
       print('Hello, service filter world!');
       return true;
     };